+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - #25 Add Patient ID in samples
+- #24 Migrate patient objects to be folderish
 - #23 Integrate new DX date field and widget
 - #18 Unique patient ID
 - #19 Convert DoB widget to native date input field